auto work 4/5 rename autos
